pyproject.toml
[tool.poetry]
name = "gp_bot"
version = "1.0.0"
description = "Gomi-People bot for Twitter"

license = "MIT"

authors = [
    "cielo <cielo.fox@gmail.com>"
]

readme = "README.md"

repository = "https://github.com/NavirePlus/gp_bot"
homepage = "https://github.com/NavirePlus"

keywords = ["Gomi-People", "ｺﾞﾐﾋﾟｰﾌﾟﾙ"]

[tool.poetry.dependencies]
PyYAML = { version = "^5.3" }
fastapi = { version = "^0.50" }
numpy = { version = "^1.18" }
python = ">=3.7 <4.0"
python-twitter = { version = "^3.5" }
torch = { version = "^1.4" }
tqdm = { version = "^4.43" }
uvicorn = { version = "^0.11" }

[tool.poetry.dev-dependencies]
autopep8 = "^1.5"
flake8 = "^3.7"
flake8-builtins = "^1.4"
flake8-docstrings = "^1.5"
flake8-isort = "^2.8"
isort = "^4.3"
mypy = "^0.761"
pip-licenses = "^2.1"
pytest = "^5.3"
pytest-cov = "^2.8"
tox = "^3.14"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
